Minor code changes and renames
