+++ src/main/java/com/jcloisterzone/game/capability/BuilderState.java
@@ -3,5 +3,5 @@ package com.jcloisterzone.game.capability;
 public enum BuilderState {
     UNUSED,
     USED,
-    SECOND_TURN;
+    SECOND_TURN
 }
