+++ src/main/java/com/jcloisterzone/game/capability/CastleCapability.java
@@ -4,7 +4,6 @@ import static com.jcloisterzone.XMLUtils.attributeBoolValue;
 
 import org.w3c.dom.Element;
 
-import com.jcloisterzone.Player;
 import com.jcloisterzone.board.Position;
 import com.jcloisterzone.feature.Castle;
 import com.jcloisterzone.feature.City;
