+++ src/main/java/com/jcloisterzone/game/capability/DragonCapability.java
@@ -5,7 +5,6 @@ import org.w3c.dom.Element;
 import com.jcloisterzone.Immutable;
 import com.jcloisterzone.board.Position;
 import com.jcloisterzone.board.TileDefinition;
-import com.jcloisterzone.board.TileGroupState;
 import com.jcloisterzone.board.TileTrigger;
 import com.jcloisterzone.figure.neutral.Dragon;
 import com.jcloisterzone.game.Capability;
