+++ src/main/java/com/jcloisterzone/game/state/MemoizedValue.java
